Made several changes to XYZs method, and replaced the old bubbles method
