+++ plotter/bubbles_test.go
@@ -7,46 +7,22 @@ package plotter
 import (
 	"image/color"
 	"log"
+	"math"
 	"math/rand"
+	"os"
 	"testing"
 
 	"gonum.org/v1/plot"
 	"gonum.org/v1/plot/internal/cmpimg"
 	"gonum.org/v1/plot/vg"
+	"gonum.org/v1/plot/vg/draw"
+	"gonum.org/v1/plot/vg/vgimg"
 )
 
-func TestBubblesRadius(t *testing.T) {
-	b := &Bubbles{
-		MinRadius: vg.Length(0),
-		MaxRadius: vg.Length(1),
-	}
-
-	tests := []struct {
-		minz, maxz, z float64
-		r             vg.Length
-	}{
-		{0, 0, 0, vg.Length(0.5)},
-		{1, 1, 1, vg.Length(0.5)},
-		{0, 1, 0, vg.Length(0)},
-		{0, 1, 1, vg.Length(1)},
-		{0, 1, 0.5, vg.Length(0.5)},
-		{0, 2, 1, vg.Length(0.5)},
-		{0, 4, 0, vg.Length(0)},
-		{0, 4, 1, vg.Length(0.25)},
-		{0, 4, 2, vg.Length(0.5)},
-		{0, 4, 3, vg.Length(0.75)},
-		{0, 4, 4, vg.Length(1)},
-	}
-
-	for _, test := range tests {
-		b.MinZ, b.MaxZ = test.minz, test.maxz
-		if r := b.radius(test.z); r != test.r {
-			t.Errorf("Got incorrect radius (%g) on %v", r, test)
-		}
-	}
-}
-
-func ExampleBubbles() {
+// ExampleScatter_bubbles draws some scatter points.
+// Each point is plotted with a different radius size depending on
+// some external criteria.
+func ExampleScatter_bubbles() {
 	rnd := rand.New(rand.NewSource(1))
 
 	// randomTriples returns some random x, y, z triples
@@ -66,29 +42,61 @@ func ExampleBubbles() {
 	}
 
 	n := 10
-	bubbleData := randomTriples(n)
+	scatterData := randomTriples(n)
+
+	// Calculate the range of Z values.
+	minZ, maxZ := math.Inf(1), math.Inf(-1)
+	for _, xyz := range scatterData {
+		if xyz.Z > maxZ {
+			maxZ = xyz.Z
+		}
+		if xyz.Z < minZ {
+			minZ = xyz.Z
+		}
+	}
 
 	p, err := plot.New()
 	if err != nil {
 		log.Panic(err)
 	}
-	p.Title.Text = "Bubbles"
+	p.Title.Text = "Bubble Plot"
 	p.X.Label.Text = "X"
 	p.Y.Label.Text = "Y"
 
-	bs, err := NewBubbles(bubbleData, vg.Points(1), vg.Points(20))
+	sc, err := NewScatter(scatterData)
 	if err != nil {
 		log.Panic(err)
 	}
-	bs.Color = color.RGBA{R: 196, B: 128, A: 255}
-	p.Add(bs)
 
-	err = p.Save(200, 200, "testdata/bubbles.png")
+	// Specify style for individual points.
+	sc.GlyphStyleFunc = func(i int) draw.GlyphStyle {
+		c := color.RGBA{R: 196, B: 128, A: 255}
+		var minRadius, maxRadius = vg.Points(1), vg.Points(20)
+		rng := maxRadius - minRadius
+		d := (float64(i)*0.8 - minZ) / (maxZ - minZ) // float64(i)*0.8 is just a random function (i)
+		r := vg.Length(d)*rng + minRadius
+		//r := float64(i)
+		return draw.GlyphStyle{Color: c, Radius: r, Shape: draw.CircleGlyph{}}
+	}
+	p.Add(sc)
+
+	img := vgimg.New(200, 200)
+	dc := draw.New(img)
+	p.Draw(dc)
+	w, err := os.Create("testdata/bubbles.png")
+	defer w.Close()
 	if err != nil {
+		log.Panic(err)
+	}
+	png := vgimg.PngCanvas{Canvas: img}
+	if _, err = png.WriteTo(w); err != nil {
+		log.Panic(err)
+	}
+	if err = w.Close(); err != nil {
 		log.Panic(err)
 	}
 }
 
-func TestBubbles(t *testing.T) {
-	cmpimg.CheckPlot(ExampleBubbles, t, "bubbles.png")
+func TestNewBubbles(t *testing.T) {
+	cmpimg.CheckPlot(ExampleScatter_bubbles, t, "bubbles.png")
 }
